linkedln-post-generator/requirements.txt
streamlit
langchain-community 
langchain-huggingface 
langchain-core
langchain
faiss-cpu 
tiktoken 
python-dotenv 
langchain-groq
playwright